requirements.txt
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.54
openpyxl>=3.1.2
pyyaml>=6.0.1
scipy>=1.11.0
